plotting: Move cache classes into `chia.plotting.cache`
